♻️ refactor: mockedSocket 클래스 다시 분리
- chatData mock data를 수정하는 함수를 export해서 사용함으로써 분리할 수 있도록 함
